server/requirements.txt
streamlit
google-genai>=0.2.0
Pillow
python-dotenv
boto3